// src/components/form/operationForm/types.d.ts
export interface ROLEDT {
    name?: string,
    id?: number,
    desc?: string
}
export interface ISubmitResult {
    value?: any;
    error: boolean;
    success: boolean;
}
export interface FormProps {
    initialValue?: any,
    onHandleClose: (e: any) => void,
    clickedBtn: string,
    handleFormData: (e: any) => void,
    categoriesList: any,
    errorMessage?: any,
    setErrorMessage?: any
}